+++ src/type.d.ts
@@ -126,6 +126,10 @@ export type ICheckbox = string | boolean | ICheckboxCof
 interface IOptionsCof {
   value: unknown
   label: string
+  icon?: string
+  tip?: string
+  viewBox?: string
+  style?: ICSSO
   disabled?: boolean
 }
 interface ISelectItemCof extends Irender {
@@ -136,6 +140,13 @@ interface ISelectItemCof extends Irender {
   changeable?: boolean
 }
 export type ISelectItem = ISelectItemCof
+
+interface IOptionGroupCof extends Irender {
+  value: unknown
+  name: string
+  options: IOptionsCof[]
+}
+export type IOptionGroup = IOptionGroupCof
 
 interface IInputItemCof extends Irender {
   value: unknown | string
@@ -218,6 +229,12 @@ interface ICodeCof extends Irender {
 }
 export type ICode = string | ICodeCof
 
+interface IColorPickerCof extends Irender {
+  name: string
+  value: unknown | string
+}
+export type IColorPicker = IColorPickerCof
+
 export type IMatchCof =
   | string
   | boolean
@@ -229,6 +246,7 @@ export type IMatchCof =
   | ISwitchItem
   | ICheckbox
   | ISelectItem
+  | IOptionGroup
   | IInputItem
   | ICanvas
   | ISliderItem
@@ -241,6 +259,7 @@ export type IMatchCof =
   | IUploadItem
   | IIFstructContainer
   | ICode
+  | IColorPicker
 
 export interface IStruct {
   tag: string