+++ CHANGELOG.md
@@ -5,3 +5,8 @@ All notable changes to the "iofod-sdk" project will be documented in this file.
 ## [1.0.0] - 2022-07-19
 
 - Initial release
+
+## [1.0.5] - 2022-08-12
+
+- Add ColorPicker component
+- Add OptionGroup component